Don't summarise RANMETHOD as an estimation method
If vou had a RANMETHOD option on a separate row (making it the first occurence of a "METH regexp-match on that row), xpose would treat it as an estimation method (eg method = 3s2p). With the updated regexp, xpose now matches only "METH" strings that are not preceded by "RAN", thus solving the issue.